Brought exceptions into line with Spring Batch standards and made them richer where feasible.
Refactored code to ensure source-factory-ref is used if configured.

Introduced notion of discarding; allows a thread to declare a reader as
finished/unusable. Addresses the scenario where a parse exception is
encountered in a non-splitting scenario and hence any further data in
the reader cannot be consumed (as we're halfway through a message with
no idea where it ends)

BufferedReaderSources now return a name (usually of the underlying
source, e.g. filename). Allows more useful exception messages.